reorg panther_zendesk_helpers
