+++ data_models/zendesk_data_model.py
@@ -1,5 +1,5 @@
 import panther_event_type_helpers as event_type
-from panther_base_helpers import ZENDESK_CHANGE_DESCRIPTION, zendesk_get_roles
+from panther_zendesk_helpers import ZENDESK_CHANGE_DESCRIPTION, zendesk_get_roles
 
 ZENDESK_TWO_FACTOR_SOURCES = {
     "Two-Factor authentication for all admins and agents",
