+++ global_helpers/panther_zendesk_helpers.yml
@@ -1,0 +1,5 @@
+AnalysisType: global
+Filename: panther_zendesk_helpers.py
+GlobalID: "panther_zendesk_helpers"
+Description: >
+  Used to define global helpers and variables for ZenDesk events.
