+++ rules/zendesk_rules/zendesk_user_suspension.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import ZENDESK_CHANGE_DESCRIPTION
+from panther_zendesk_helpers import ZENDESK_CHANGE_DESCRIPTION
 
 USER_SUSPENSION_ACTIONS = {
     "create",
